requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-core==1.21.0
azure-identity==1.7.1
azure-purview-scanning==1.0.0b2
azure-purview-catalog==1.0.0b2
